refactor(migration-fix-wizard): 1472줄 갓파일을 도메인 5모듈로 분할 (#62)
migration_fix_wizard.py(1472줄)를 순수 동작 보존 리팩터로 5개 도메인 모듈에
물리 분할하고, 원본은 얇은 facade(__all__ 14개 re-export)로 남겨 9개 소비자의
import 경로를 무변경 유지한다.

- [CC-065] SmartFixGenerator+create_wizard_steps→migration_fix_option_generator,
  CollationFKGraphBuilder→migration_fk_graph, FKSafeCharsetChanger→
  migration_fk_safe_charset, BatchFixExecutor→migration_batch_fix_executor,
  CharsetFixPlanBuilder→migration_charset_fix_plan. import DAG는 fk_graph를 leaf로
  순환 없이 구성.
- [CC-068] 4개 클래스 verbatim lazy-init을 migration_fk_graph.build_fk_graph
  헬퍼로 추출, per-instance 캐시 가드만 유지.
- [CC-066/067/074] execute_batch를 _execute_fk_safe_clusters/
  _execute_collation_single_merges/_execute_remaining_steps 3개로 분해,
  execute_batch는 concat+집계만 담당. 빈 rollback_sql 지역변수 제거(리터럴 인라인),
  클래스 docstring을 dry-run 전용으로 재작성.
- [CC-075] module-level logger 추가, FK 정렬/sql_mode 복원 실패 시 진단 로그.
- [CC-069] invalid-date WHERE 절을 _invalid_date_where_clause 헬퍼로 추출.
- [CC-072] DEFAULT_TARGET_CHARSET/COLLATION 상수를 models에 1회 정의 후 리터럴 교체.
- [CC-070] INFORMATION_SCHEMA.TABLES charset 조회를 models.get_table_charset
  공유 함수로 추출(read-only).
- [CC-073] FixWizardStep의 미사용 included_by/included_reason 제거 + 도달 불가
  롤백 분기/obsolete assert 삭제.
- [CC-081] generate_rollback_sql를 전략별 헬퍼(_rollback_date/
  _rollback_collation_single/_rollback_collation_fk)로 분해.

동작 보존: public 시그니처/출력 문자열/dry_run=False fail-closed 가드 유지.
검증: py_compile 8파일, 타겟 pytest 209 pass, 전체 pytest 1836 pass
(macOS CI 의존 flaky 1건 제외).
